Atualização para incluir o período 2020.3

+++ core/visoes/enquete_view.py
@@ -1,3 +1,6 @@
+from datetime import datetime
+
+import pytz
 from django.contrib import messages
 from django.contrib.auth.decorators import permission_required
 from django.shortcuts import render, redirect, get_object_or_404
@@ -15,6 +18,11 @@ def enquete_voto_view(request, pk, abstencao=None):
     discente = check_discente(request)
 
     enquete = Enquete.objects.get(pk=pk)
+
+    url_redirect = check_periodo_enquete(request, enquete)
+    if url_redirect:
+        return url_redirect
+
     mesmo_curso = check_curso(enquete, discente)
     if not mesmo_curso:
         messages.error(request, 'Você não pode votar na enquete "' + enquete.nome + '".')
@@ -53,6 +61,24 @@ def enquete_voto_view(request, pk, abstencao=None):
         'votos_por_discente': votos_por_discente,
     }
     return render(request, 'core/enquetes/votar.html', context)
+
+
+def enquete_detalhar_voto_view(request, pk, abstencao=None):
+    usuario = request.user
+    if not discente_existe(usuario):
+        messages.error(request, 'Não há um discente relacionado ao usuário.')
+        return redirecionar(request)
+    discente = check_discente(request)
+    enquete = Enquete.objects.get(pk=pk)
+
+    qtd_votos = get_qtd_votos(enquete, discente)
+    votos_por_discente = get_votos(enquete, discente)
+    context = {
+        'enquete': enquete,
+        'qtd_votos': qtd_votos,
+        'votos_por_discente': votos_por_discente,
+    }
+    return render(request, 'core/enquetes/detalhar_voto.html', context)
 
 
 def tem_componente(form_voto):
@@ -128,6 +154,9 @@ def enquete_deletar_voto_discente(request, pk, template_name='core/enquetes/voto
     if not tem_permissao(request, voto_turma.discente):
         messages.error(request, 'Você não tem permissão de Excluir este voto.')
         return redirecionar(request)
+    url_redirect = check_periodo_enquete(request, voto_turma.enquete)
+    if url_redirect:
+        return url_redirect
     if request.method == 'POST':
         voto_turma.delete()
         messages.success(request, 'Voto na enquete excluído com sucesso.')
@@ -172,3 +201,13 @@ def load_componente(request):
     componente_id = request.GET.get('componenteId')
     componente = ComponenteCurricular.objects.get(pk=componente_id)
     return render(request, 'core/enquetes/requisitos_list.html', {'componente': componente})
+
+
+def check_periodo_enquete(request, enquete):
+    agora = datetime.now()
+    utc = pytz.UTC
+    if enquete.data_hora_inicio.replace(tzinfo=utc) > agora.replace(tzinfo=utc) \
+       or agora.replace(tzinfo=utc) > enquete.data_hora_fim.replace(tzinfo=utc):
+        messages.error(request, 'Voto fora do período da Enquete "' + enquete.nome + '".')
+        return redirect('/core/enquetes/list')
+    return None
